refactor: components 파일 구조 수정
