+++ src/components/chatting/ChattingProfile.tsx
@@ -1,10 +1,10 @@
-import { ReactComponent as Back } from "../assets/icons/Back.svg";
-import { ReactComponent as Call } from "../assets/icons/Call.svg";
-import { ReactComponent as Friends } from "../assets/icons/Freinds.svg";
-import { ReactComponent as More } from "../assets/icons/More.svg";
-import { Heading2, Caption1 } from "../style/font";
-import "../style/color.css";
-import { usersState } from "../recoil/state";
+import { ReactComponent as Back } from "../../assets/icons/Back.svg";
+import { ReactComponent as Call } from "../../assets/icons/Call.svg";
+import { ReactComponent as Friends } from "../../assets/icons/Freinds.svg";
+import { ReactComponent as More } from "../../assets/icons/More.svg";
+import { Heading2, Caption1 } from "../../style/font";
+import "../../style/color.css";
+import { usersState } from "../../recoil/state";
 import { useRecoilValue } from "recoil";
 import styled from "styled-components";
 import { useNavigate } from "react-router-dom";
