+++ src/data/chatData.json
@@ -8,13 +8,15 @@
           "chatId": 0,
           "userId": 0,
           "message": "에어팟을 끼고 일해도 힘듭니다",
-          "date": "Thu Sep 28 2023 09:41:56 GMT+0900 (한국 표준시)"
+          "date": "Thu Sep 28 2023 09:41:56 GMT+0900 (한국 표준시)",
+          "like": 0
         },
         {
           "chatId": 1,
           "userId": 3,
           "message": "앗 그렇군뇽 그냥 하세요",
-          "date": "Thu Sep 28 2023 10:31:42 GMT+0900 (한국 표준시)"
+          "date": "Thu Sep 28 2023 10:31:42 GMT+0900 (한국 표준시)",
+          "like": 0
         }
       ]
     },
@@ -26,13 +28,15 @@
           "chatId": 0,
           "userId": 0,
           "message": "배고파요",
-          "date": "Wed Oct 4 2023 16:18:14 GMT+0900 (한국 표준시)"
+          "date": "Wed Oct 4 2023 16:18:14 GMT+0900 (한국 표준시)",
+          "like": 0
         },
         {
           "chatId": 1,
           "userId": 2,
           "message": "네",
-          "date": "Wed Oct 4 2023 16:18:16 GMT+0900 (한국 표준시)"
+          "date": "Wed Oct 4 2023 16:18:16 GMT+0900 (한국 표준시)",
+          "like": 0
         }
       ]
     },
