+++ src/pages/ChattingPage.tsx
@@ -1,6 +1,6 @@
-import ChattingRoom from "../components/ChattingRoom";
-import ChattingInput from "../components/ChattingInput";
-import ChattingProfile from "../components/ChattingProfile";
+import ChattingRoom from "../components/chatting/ChattingRoom";
+import ChattingInput from "../components/chatting/ChattingInput";
+import ChattingProfile from "../components/chatting/ChattingProfile";
 import styled from "styled-components";
 import { useState, useEffect } from "react";
 import { useRecoilState, useRecoilValue } from "recoil";
@@ -10,33 +10,29 @@ import {
   nowUserIdState,
 } from "../recoil/state";
 import { useParams } from "react-router-dom";
-interface ChattingPageProps {
-  chatRoomId: number;
-}
-function ChattingPage(props: ChattingPageProps) {
+
+function ChattingPage() {
+  let { id } = useParams();
   //chatting 전체 리스트
   const chatRoomList = useRecoilValue(chatRoomListState);
   //해당 채팅방 chatting list
   const [chatListById, setChatListById] = useRecoilState(
-    chatListByIdState(props.chatRoomId)
+    chatListByIdState(Number(id))
   );
   //채팅방 사용자 2명
-  const roomUsers = chatRoomList[props.chatRoomId].userList;
+  const roomUsers = chatRoomList[Number(id)].userList;
   //user switch 상태
   const [switchUser, setSwitchUser] = useState(0);
   //list 길이 체크해서 빈 채팅방 or 채팅있는 채팅방
-  const chatList = chatRoomList[props.chatRoomId].chatList;
+  const chatList = chatRoomList[Number(id)].chatList;
   //현재 사용중인 user의 id(login user 아님)
   const [nowUserId, setNowUserId] = useRecoilState(nowUserIdState);
   //localstorage에 없으면 json 파일 저장, 있으면 chatList 불러옴
-  const initializeChat = localStorage.getItem(
-    "chat" + String(props.chatRoomId)
-  );
-  let { id } = useParams();
+  const initializeChat = localStorage.getItem("chat" + String(Number(id)));
   useEffect(() => {
     if (initializeChat === null) {
       localStorage.setItem(
-        "chat" + String(props.chatRoomId),
+        "chat" + String(Number(id)),
         JSON.stringify(chatListById)
       );
     } else {
@@ -60,8 +56,8 @@ function ChattingPage(props: ChattingPageProps) {
         <ChattingProfile userId={nowUserId} />
       </div>
       <Divider />
-      <ChattingRoom chatRoomId={props.chatRoomId} isUser={true} />
-      <ChattingInput chatRoomId={props.chatRoomId} />
+      <ChattingRoom chatRoomId={Number(id)} isUser={true} />
+      <ChattingInput chatRoomId={Number(id)} />
     </div>
   );
 }
